refactor(server): /api/notifications 라우트 분할 (Plan C #C9 1단계)
- src/server/routes/notifications.ts 신규: 7개 라우트 캡슐화
- registerNotificationsRoutes(api, ctx) 패턴 도입
- safeError 적용 (sanitizeErrorMessage(getErrorMessage(e)) → safeError(c, e, prefix))
- dashboard-api.ts 2119 → 1977줄 (-142)

후속: jobs/projects/config/setup/stats/version/repositories/health/doctor/sse 도메인을
      동일 패턴으로 단계 분할 예정
