+++ src/server/dashboard-api.ts
@@ -36,6 +36,7 @@ import { z } from "zod";
 import type { ZodError } from "zod";
 import { loadTemplate, renderTemplate } from "../prompt/template-renderer.js";
 import { SSEManager } from "./sse-manager.js";
+import { registerNotificationsRoutes } from "./routes/notifications.js";
 
 // Session manager: in-memory token store with TTL and periodic pruning
 const sessionManager = new SessionManager();
@@ -1969,151 +1970,8 @@ export function createDashboardRoutes(store: JobStore, queue: JobQueue, configWa
     });
   });
 
-  // Notifications: unread count
-  api.get("/api/notifications/unread-count", (c) => {
-    try {
-      const unreadCount = store.getAqDb().countUnreadNotifications();
-      return c.json({ unreadCount });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to fetch unread count: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: list with pagination
-  api.get("/api/notifications", (c) => {
-    try {
-      const queryParams = {
-        isRead: c.req.query("isRead"),
-        type: c.req.query("type"),
-        limit: c.req.query("limit") ? parseInt(c.req.query("limit")!, 10) : undefined,
-        offset: c.req.query("offset") ? parseInt(c.req.query("offset")!, 10) : undefined,
-      };
-
-      const parseResult = GetNotificationsQuerySchema.safeParse(queryParams);
-      if (!parseResult.success) {
-        return c.json({
-          error: "Invalid query parameters",
-          details: formatZodError(parseResult.error)
-        }, 400);
-      }
-
-      const { isRead, type: typeFilter, limit, offset } = parseResult.data;
-      const aqDb = store.getAqDb();
-      const isReadFilter = isRead === "true" ? true : isRead === "false" ? false : undefined;
-      const notifFilter: { isRead?: boolean; type?: string } = {};
-      if (isReadFilter !== undefined) notifFilter.isRead = isReadFilter;
-      if (typeFilter !== undefined) notifFilter.type = typeFilter;
-      const notifications = aqDb.listNotifications({ ...notifFilter, limit, offset });
-      const total = aqDb.countNotifications(notifFilter);
-      const unreadCount = aqDb.countUnreadNotifications();
-      const start = offset ?? 0;
-
-      return c.json({
-        notifications,
-        total,
-        unreadCount,
-        pagination: {
-          total,
-          offset: start,
-          limit: limit ?? total,
-          hasMore: start + notifications.length < total,
-        },
-      });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to fetch notifications: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: mark all as read (static route before :id/read)
-  api.post("/api/notifications/read-all", (c) => {
-    try {
-      const count = store.getAqDb().markAllNotificationsRead();
-      broadcastToAllClients("notificationsReadAll", { count, timestamp: Date.now() });
-      return c.json({ status: "ok", count });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to mark all notifications as read: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: mark individual as read
-  api.post("/api/notifications/:id/read", (c) => {
-    try {
-      const idParam = c.req.param("id");
-      const id = parseInt(idParam, 10);
-      if (isNaN(id) || id <= 0) {
-        return c.json({ error: "Invalid notification id" }, 400);
-      }
-      const updated = store.getAqDb().markNotificationRead(id);
-      if (!updated) {
-        return c.json({ error: "Notification not found" }, 404);
-      }
-      return c.json({ status: "ok", id });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to mark notification as read: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: prune read notifications older than maxAgeDays (default 14)
-  api.post("/api/notifications/prune", async (c) => {
-    if (readOnly) {
-      return c.json({ error: "Read-only mode" }, 403);
-    }
-    try {
-      let maxAgeDays = 14;
-      try {
-        const body = await c.req.json<{ maxAgeDays?: unknown }>();
-        if (typeof body?.maxAgeDays === "number" && body.maxAgeDays > 0) {
-          maxAgeDays = Math.floor(body.maxAgeDays);
-        }
-      } catch {
-        // body 없음 → 기본값 14일
-      }
-      const deleted = store.pruneReadNotifications(maxAgeDays);
-      broadcastToAllClients("notificationsPruned", { deleted, maxAgeDays, timestamp: Date.now() });
-      return c.json({ deleted, maxAgeDays });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to prune notifications: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: delete all (옵션으로 isRead 필터)
-  api.delete("/api/notifications", (c) => {
-    if (readOnly) {
-      return c.json({ error: "Read-only mode" }, 403);
-    }
-    try {
-      const isReadParam = c.req.query("isRead");
-      const filter: { isRead?: boolean } = {};
-      if (isReadParam === "true") filter.isRead = true;
-      else if (isReadParam === "false") filter.isRead = false;
-      const deleted = store.deleteAllNotifications(filter);
-      broadcastToAllClients("notificationsDeleted", { deleted, timestamp: Date.now() });
-      return c.json({ deleted });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to delete notifications: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
-
-  // Notifications: delete single
-  api.delete("/api/notifications/:id", (c) => {
-    if (readOnly) {
-      return c.json({ error: "Read-only mode" }, 403);
-    }
-    try {
-      const idParam = c.req.param("id");
-      const id = parseInt(idParam, 10);
-      if (isNaN(id) || id <= 0) {
-        return c.json({ error: "Invalid notification id" }, 400);
-      }
-      const deleted = store.deleteNotification(id);
-      if (!deleted) {
-        return c.json({ error: "Notification not found" }, 404);
-      }
-      return c.json({ status: "ok", id });
-    } catch (error: unknown) {
-      return c.json({ error: `Failed to delete notification: ${sanitizeErrorMessage(getErrorMessage(error))}` }, 500);
-    }
-  });
+  // Notifications: 도메인 라우트 분할 (Plan C #C9)
+  registerNotificationsRoutes(api, { store, sseManager, readOnly: !!readOnly });
 
   return api;
 }
